Change permission failure from fatal to warning.

+++ Library/Homebrew/cask/artifact/moved.rb
@@ -84,8 +84,7 @@ module Cask
 
         Utils.gain_permissions_mkpath(target.dirname, command: command) unless target.dirname.exist?
 
-        if target.directory?
-          Quarantine.ensure_app_management_permissions_granted(app: target, command: command)
+        if target.directory? && Quarantine.app_management_permissions_granted?(app: target, command: command)
           if target.writable?
             source.children.each { |child| FileUtils.move(child, target/child.basename) }
           else
@@ -154,10 +153,11 @@ module Cask
 
         return unless Utils.path_occupied?(target)
 
-        if target.directory? && matching_artifact?(successor)
+        if target.directory? && matching_artifact?(successor) && Quarantine.app_management_permissions_granted?(
+          app: target, command: command,
+        )
           # If an app folder is deleted, macOS considers the app uninstalled and removes some data.
           # Remove only the contents to handle this case.
-          Quarantine.ensure_app_management_permissions_granted(app: target, command: command)
           target.children.each do |child|
             if target.writable? && !force
               child.rmtree
